made changes to existing testcases.
