+++ tests/test_web_scrapper.py
@@ -1,0 +1,4 @@
+from ..code.web.scraper.web_scraper import scraper
+
+def test_get_driver():
+    assert scraper() is not None
